requirements.txt
djangorestframework
django-filter
markdown
Pillow
djoser